fix: remove unnecessary cap from bluetooth script
Since `/etc/modprobe.d` is writable by root, `CAP_DAC_OVERRIDE` is not
needed. Also fix formatting of `CapabilityBoundingSet` property and did
some minor code cleanup/style changes:

* Moved `SandboxedFunction` to `__init__.py`; it's few enough lines of
  code that there's no need for a separate submodule.
* Removed unreachable code: the properties of `SandboxedFunction` of
  type `list[str]` cannot be `None`.
* Added return type hints to all functions.
* Removed unnecessary type hints from local variables.
* Added shebang line to Python scripts.
