--- files/system/usr/libexec/secureblue/sandbox/sandboxed_function.py
@@ -1,29 +1,0 @@
-# Copyright 2025 The Secureblue Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-A class that wraps a function to be run in a sandbox
-"""
-
-import dataclasses
-
-
-@dataclasses.dataclass
-class SandboxedFunction:
-    """A class that wraps a function to be run in a sandbox"""
-
-    file_name: str
-    capabilities: list[str] = dataclasses.field(default_factory=list)
-    read_write_paths: list[str] = dataclasses.field(default_factory=list)
-    additional_sandbox_properties: list[str] = dataclasses.field(default_factory=list)
